fix(core): Inline config.js to index.html to prevent CF from caching it
config.js contains the FE sentry config, as well the REST endpoint's base path.
It's served from the `/rest/config.js` endpoint. We don't apply any cache
headers to responses from the `/rest` path. However, Cloudflare
[by default](https://developers.cloudflare.com/cache/concepts/default-cache-behavior/)
caches responses that have default file extension (e.g. `.js`) and don't include
the `cache-control` header.

This PR inlines the contents of `config.js` into the `index.html` file. This solves
the caching issue as well as removes one request on the hot path of app startup.
