+++ packages/cli/src/server.ts
@@ -273,22 +273,6 @@ export class Server extends AbstractServer {
 				`/${this.restEndpoint}/settings`,
 				ResponseHelper.send(async () => frontendService.getSettings()),
 			);
-
-			this.app.get(`/${this.restEndpoint}/config.js`, (_req, res) => {
-				const frontendSentryConfig = JSON.stringify({
-					dsn: this.globalConfig.sentry.frontendDsn,
-					environment: process.env.ENVIRONMENT || 'development',
-					serverName: process.env.DEPLOYMENT_NAME,
-					release: `n8n@${N8N_VERSION}`,
-				});
-				const frontendConfig = [
-					`window.REST_ENDPOINT = '${this.globalConfig.endpoints.rest}';`,
-					`window.sentry = ${frontendSentryConfig};`,
-				].join('\n');
-
-				res.type('application/javascript');
-				res.send(frontendConfig);
-			});
 		}
 
 		// ----------------------------------------
